refactor(pages): extract hotline number pill logic to HotlineNumber component
